moved modal components to its own folder
