+++ client/components/modal/confirm-modal.jsx
