+++ client/components/modal/delete-confirm.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './styles/delete-confirm.css';
+import '../styles/delete-confirm.css';
 
 export default function DeleteConfirm(props) {
   return (
